Final Edit for Transfer
Shifting project to home computer. BBCode editor doesn't work in current
state and needs to be finished editing. However, there are adequate
notes on what to do.

Remember:
HTML Preset to parse it and then stick that in created HTML to store the
state. Also, given the size of the state that I'm likely to do, this is
probably a good project to have redux on.